chore: src/ add .test format at strangly files
