+++ frontEnd/src/components/Algorithm.js
