+++ frontEnd/src/components/Architecture.js
